+++ src/accumulator.js
@@ -1,16 +1,16 @@
 module.exports = accumulator
 
-function accumulator(itemName = 'item', listName = 'list') {
+function accumulator(validator = isNotEmpty, itemName = 'item', listName = 'list') {
   const storage = {}
-  return { method, getter }
+  return { item, list }
 
-  function method(name, obj) {
+  function item(name, obj) {
     if (arguments.length > 1) {
       if (storage.hasOwnProperty(name)) {
         throw new Error(`${itemName} "${name}" already present in ${listName}`)
       }
-      if (obj == undefined) { // if empty obj
-        throw new Error(`value for ${itemName} "${name}" should not be empty`)
+      if (validator && !validator(obj)) {
+        throw new Error(`value for ${itemName} "${name}" is not valid`)
       }
       storage[name] = obj
     } else {
@@ -21,7 +21,11 @@ function accumulator(itemName = 'item', listName = 'list') {
     return storage[name]
   }
   
-  function getter() {
+  function list() {
     return Object.keys(storage)
   }
+}
+
+function isNotEmpty(obj) {
+  return (obj != undefined)
 }