+++ CHANGELOG.md
@@ -1,19 +1,23 @@
-## 0.1.4 / 2019-05-17
+## 0.2.2 / 2019-05-17
 - update `README.md`
 - add build (and minified version) ready for the browser
 
 
-## 0.1.3 / 2019-05-13
+## 0.2.1 / 2019-05-13
 - update `README.md`
 - add usage examples
 - do some refactoring
 - publish on npm@v0.1.3
 
 
-## 0.1.2 / 2019-05-10
+## 0.2.0 / 2019-05-10
 - update `README.md`
+- define-accumulator.test refactoring 
 - add validation function as a parameter in order to check
   that objects of the same type are being added (can use: instanseof, duck typing, validation...)
+- set as a default validation the restriction: cannot add empty object in storage
+- test refactoring
+- rename functions `method` and `getter` to `item` and `list` accordingly
 
 ## 0.1.1 / 2019-05-08
 - update `README.md`